Add support for multiple feeds
This refactors the JS to SpotGroup and spot_feed.display to decide
what is mapped.

It also adds custom pins, has an early form of color/per/feed and adds
pin clustering for better performance.

I still need to figure out something better for colors. They need to be
more contrasting but automated.